pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-ssh-gateway"
version = "1.0.0"
description = "Remote MCP Server for ChatGPT with SSH Gateway to VPS"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MCP Gateway Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=2.0.0",
    "httpx>=0.26.0",
    "httpx-sse>=0.3.0",
    "paramiko>=3.4.0",
    "asyncssh>=2.14.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "redis>=5.0.0",
    "slowapi>=0.1.9",
    "structlog>=24.1.0",
    "pyyaml>=6.0.1",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
mcp-server = "mcp_server.main:run_server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
